Update patch set 1
Patch Set 1: Code-Review-2

This does not work on most devices. It expects keymaster HAL services to set this property but only a fraction of devices use the keymaster service that you modified. In-fact I am sure that using a property for communication across the system-vendor boundary is a treble violation.

Patch-set: 1
Reviewer: Gerrit User 1092085 <1092085@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2