Update patch set 3
Patch Set 3:

(3 comments)

Sorry I haven't gotten back to this.  I've been thinking about this and other potential options here and I'm still not coming up with any clear best choices.

The code that sorts the messages by realtime fundamentally breaks any plans that try to use timestamp or even sequence numbers to make the 'region lock'.  That code likely needs to go if we continue with this.

I was assuming that we only had one thread writing the log messages, which makes the monotonic clock also a sequence number, but I now realize that the audit log and the kernel log come from different threads, so that may also create a situation where we have out of order timestamps.

I've been thinking of changing flushTo() to take an return iterators instead.  This may be the safest option.  We'd want to have a std::optional<LogBufferElementCollection::iterator> as the input.  If it's nullopt, then we start at the beginning of the list, otherwise we start flushing at the next element.  We then return the last element that we've printed.  (Since we cannot return the 'end' iterator then use it again).

That approach should be totally safe in terms of knowing where each thread is reading from, since we'd literally know which iterators they're interested in.  We'd need to do more work to convert the 'start' timestamp into this iterator however.  I think that's the best option here.

Thanks for your help with this.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>