Update patch set 3
Patch Set 3:

Discussion still makes sense to be public.

> Patch Set 3:
> There are vdso, at least on 64bit, now though, right?  It shouldn't be a big performance cost anymore, and the log will be much more accurate if we can sort by monotonic_raw time.  We'd also be able to do more accurate kernel and userspace log merging, which is a strong benefit.

monotonic_raw and boottime were not part of the list of supported vDSO clocks in the past. arm32 and its lack of vDSO32 still figured prominently too. For instance the audio server had a big complaint about the latency for logging that we tried to address by reducing what we collected. We put the heat up on vDSO32, which is in now after 4 years of pushing against multiple stakeholders. These should all be by the wayside improving the picture and why _today_ I would like to see monotonic_raw, realtime and bottime collected.

Is audio server still 32-bit? Hopefully not ...

> What does the boottime clock get us, besides the nice FYI of how long since boot did this take place?

Track suspend and hibernate accurately. And logd can be greatly simplified and made easier to maintain.

Upstream kernel has resisted any suspend or hibernate tracking and logging with spot messages with a specific signature reporting all the timestamps. If we add boottime collection we can remove yet another out-of-kernel patch and maybe a whole bunch of code in the logger that sniffs for the kernel logs that handle sleeping. There has also been some requests for boottime in the past that I pushed back on because it was so costly to collect and to this day I do not know how I can easily differentiate from boottime and monotonic_raw algorithmically (there is infrastructure in logger that needs to be replaced that relies on determining if a timestamp is monotonic by checking if it is ridiculously old; a less heuristic algorithm would be appreciated).

Interestingly the kernel folks that resisted adding sleep logging countered by adding realtime and boottime to the internal kernel log buffer monotonic_raw recording API. So the idea for this came from their feedback. That patch never made it in because Linus countered by replacing monotonic_raw with boottime for dmesg logging, and it broke userspace. The bruise from that effort seemed to be lasting as the author of the time-triple lost his nerve, as did I. kernel logging is back to "it works, don't futz with it" ...

The time triple (estimating boottime and realtime) will still be fraught with difficulties in logd.klogd and logd.auditd. All solved if the logger _only_ works in montonic_raw time with the loss of suspend tracking though.

> > > I've been thinking of changing flushTo() to take an return iterators instead.
> > 
> > Prune can cause an iterator to go away, and with heavy load from the spam filter from a large swath in the middle ... already tried (something like) that and got the surprise of a lifetime. Maybe we should invent a 'safe iterators' class that can get notification of context loss and skips to the next entry?
> 
> Prune will only erase iterators after the watermark though, right?  The idea here would be to use actual iterators as the watermark and then have flushTo() take and return the 'last read' iterator.

Yes, true, that lower watermark protection was added to help address this and I did not get back to trying again.

+2 on Iterators would make us time-independent (except the initial search for tail by time).

Patch-set: 3
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>