Update patch set 5
Patch Set 5:

You probably want to move this back into Action::AddCommand() to catch when people use this as a command outside of service restarts.

There haven't been many cases, but this may make it worth creating a secondary set of functions that splits the validation of the arguments of builtins from the actual action itself.  For example:

static Result<void> do_setrlimit(const BuiltinArguments& args) {
    auto rlimit = ParseRlimit(args.args);
    if (!rlimit) return rlimit.error();

    if (setrlimit(rlimit->first, &rlimit->second) == -1) {
        return ErrnoError() << "setrlimit failed";
    }
    return {};
}

Could be split into a function where ParseRlimit() is called, which would also be callable in host_init_verifier, and then a separate function that just uses the result for the actual setrlimit call.

Similarly here, you'd have a function that does this check, then a separate function that actually does the action.  The former could be run in host_init_verifier.

The only thing that does break a lot of this, is there's nothing stopping someone from doing interface_start ${some.property}, which we can't check on host at all.

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>