Update patch set 2
Patch Set 2: Autosubmit+1

> Patch Set 2: Code-Review+2
> 
> oops! sorry about that.
> 
> (this isn't a _recent_ change though --- it was back in January. are we missing a test that would have helped find this earlier?)
> 
> i've sent a patch upstream to stop silently ignoring -I at least...

Thanks for your review.

Yes there was no test to check this behavior. As long as we are moving this to runtime generator, I think we can add more tests on it.

Also, I mentioned recent quite relatively (compared to when this was first implemented). Feels sad that we failed to find this in earlier.

Patch-set: 2
Reviewer: Gerrit User 1413436 <1413436@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=+1
Label: Presubmit-Ready=0