Enable "localfilesystem" UNIX domain socket for ADB.
This patch introduce "service.adb.listen_addr", a new
string type property, while keeping the old properties.
The new property added in this patch is used to listen
on UNIX domain socket "localfilesystem".

In the process of introducing the new socket type, the
method tcp_connect is removed and combined into
socket_spec_connect.

Some examples of the new property value are:
  - "tcp:5555"
  - "vsock:5555"
  - "localfilesystem:/tmp/test"

Bug: 133378083
Test: On master-arc-dev:
	adb root;
	setprop service.adb.listen_addr <path_to_socket>;
	adb connect <path_to_socket>;
      	adb -s <path_to_socket> shell; inside Chrome OS.
Test: On aosp_bluefin:
	setprop service.adb.listen_addr tcp:5555;
	adb connect <IP>:5555; adb shell;
Test: On aosp_bluefin:
	setprop service.adb.tcp.port 5555;
	adb connect <IP>:5555; adb shell;

Change-Id: I8289bf0ab3305cf23ce5695ffba46845d58ef6bb