Update patch set 3
Patch Set 3:

> Patch Set 3:
> . . .
> > > What does the boottime clock get us, besides the nice FYI of how long since boot did this take place?
> > 
> > Track suspend and hibernate accurately. And logd can be greatly simplified and made easier to maintain.
> > 
> > Upstream kernel has resisted any suspend or hibernate tracking and logging with spot messages with a specific signature reporting all the timestamps. If we add boottime collection we can remove yet another out-of-kernel patch and maybe a whole bunch of code in the logger that sniffs for the kernel logs that handle sleeping. There has also been some requests for boottime in the past that I pushed back on because it was so costly to collect and to this day I do not know how I can easily differentiate from boottime and monotonic_raw algorithmically (there is infrastructure in logger that needs to be replaced that relies on determining if a timestamp is monotonic by checking if it is ridiculously old; a less heuristic algorithm would be appreciated).
> 
> We do all of this to guess at what the realtime time stamp is for kernel logs, right?

Yes to give us the ability to merge kernel klogd, auditd and user space primarily, heuristics and hacks I was never comfortable with but had to do to keep everyone happy. Admittedly with about 3 ms of peak error at times which gives us loads of troubles with avc and kernel log alignment at times. The error goes to nanosecond when we switch the logger to monotonic (Android Wear and Android TV I think do this), and I expect the same if we switch to doubles or triples for the timestamp(s) and propagate them out the logger-pipes to the readers.

But (outside of logd) that message and values are used to track suspend time. If we go triple (or even double, but boottime is _much_ better for this than realtime because of updates from vendor time sources creating discontinuities and hence why I advocate triples), we should be able to drop all this and once done, approach the folks that track suspend time in the UI and beyond to use the logger timing data to determine suspend time, and _then_ we can remove the out-of-tree patches in the kernel to help track suspend (and hibernate) times and enjoy a future of less maintenance all around.

NB: logger is not an API because of CPU load and DOS attack, alternate/independent means of tracking suspend is more architecturally sound. Hopefully they have abandoned the logger and recognized this. I have no visibility into how they do this today.

monotonic_raw == kernel and avc log time source, no rate correction.
boottime == rate corrected time with continuity
realtime == rate and _time_ corrected with major discontinuities, but humans can relate to better.

NB: auditd logs are _not_ aligned with the denials, the timestamps in message and in metadata are not the same as they are processed after the fact from a queue. Sometimes 100's of us later. Would love to find a solution to make the events closer to the activities whether they be through (ugly) logd heuristics, or better timestamp regime in the kernel.

. . .
> > +2 on Iterators would make us time-independent (except the initial search for tail by time).
> 
> I'm going to try this (once again).  The tail time to iterator check is a little bit rough, but doable.

sgtm, exciting times, a long time(sic) coming!

Patch-set: 3
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>