Update patch set 11
Patch Set 11:

Since we are going LMKD diet to make it fast, this CL goes opposite way, IMO.
If there is something we should read 4KB above, that gives me a signal "we should fix something", rather than allowing it.

zoneinfo, kernel already went node LRU instead of zone's one.
IOW, zoneinfo would be less important than old day by day.
Why do we need to read zoneinfo?
Only reason at this moment is get to totalreserve_pages which is add up of protection pages of each zone.
They are almost one time setup values at booting and updated with lowmem_reserve_ratio.
It would be very rare to change on demand(Actually, with my short experience, I never have seen to change it in runtime).
If someone want to really change it on runtime, can't we create some API so that LMKD only get the message when it is updated. With that, LMKD doesn't need to re-read the same value all the time.

About the status, it is used by only to get tgid. Then, are't we enough to read 1Kbyte from the status?

Patch-set: 11
Reviewer: Gerrit User 1275554 <1275554@85c56323-6fa9-3386-8a01-6480fb634889>