Update patch set 1
Patch Set 1:

You need this to be able to actually try to connect to each of the interfaces, right?

I remember having not wanted to export the PIDs of services in the past, but I can't think of the reason why right now.

+nnk@ if there's a security reason that I'm forgetting.

In any case, if we do expose the PIDs, then you should have an SEPolicy that prohibits them from being read outside of the processes that absolutely need it (tests).  Since init.svc.* is already exposed, probably better if these properties were init.svc_pid.* as well.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>