Update patch set 4
Patch Set 4:

PS4: I change to link libcharger_sysprop statically. 

(1) The library is only used by charger (in system image) / charger.recovery (in recovery / boot image) and not reused, so it is the best to be linked statically. 

(2) charger.recovery is not marked as a recovery module, so libcharger_sysprop.so isn't installed to the recovery image, causing the TH failure.

Patch-set: 4
Reviewer: Gerrit User 1121838 <1121838@85c56323-6fa9-3386-8a01-6480fb634889>