Update patch set 1
Patch Set 1:

(3 comments)

Sounds good to me.  Thanks for debugging that CF issue.  I was hoping the other crash handling code I put in place would cover all cases, but it looks like it doesn't.

This does mean that anything that gets fork'ed from selinux_setup (and we do fork the SELinux compiler in the case that the precompiled sepolicy hash doesn't match) will inherit these FDs.  I'm not sure if that's a feature or a bug.  We can prevent that by calling SetStdioToDevNull() again in the beginning of SetupSelinux().

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>