Update patch set 2
Patch Set 2: Presubmit-Ready+1

Yes, wanted it to also deal with second stage init exec as a side effect, but for this bug's specific purposes it is the transition from ramdisk in first stage init, switch root, on a 4.19 kernel that is unique and probably triggered the problem(s) and I did not notice a similar problem for second stage init. I would want to make sure we have no uncontrolled fds through any exec or fork transitions and if a problem does show up we can rubberstamp this into the next exec transition on an as-need basis.

(bug is still under investigation and prioritization)

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1