Adding footer in super image123
This commit adds the footer in super image to eliminate the
need of adding /vbmeta_system and/or /vbmeta_vendor when AVB
chain partition is used with Android Dynamic Partition.

See BOARD_AVB_VBMETA_SYSTEM under the link:
https://android.googlesource.com/platform/external/avb/#build-system-integration

|Backup Super AVB Footer| (variable-length)
|    Super AVB Footer   | (variable-length)
|  Backup Super Footer  | (fixed-length)
|      Super Footer     | (fixed-length)

The bottom Super Footer records the physical offset of the Super
AVB Footer, which in turn records the physical offset of each
logical partition's vbmeta within the /super partition.

Bug: 137054296
Test: m liblp_test_host
Test: ./out/host/linux-x86/nativetest64/liblp_test_host/liblp_test_host
Test: m libvbmeta_test
Test: ./out/host/linux-x86/nativetest/libvbmeta_test/libvbmeta_test
Change-Id: I01aeadd850750ae87d9125484c1b1f570bb84756