Construct the super_vbmeta image
This commit constructs the super_vbmeta image to eliminate the
need of adding /vbmeta_system and/or /vbmeta_vendor when AVB
chain partition is used with Android Dynamic Partition.

See BOARD_AVB_VBMETA_SYSTEM under the link:
https://android.googlesource.com/platform/external/avb/#build-system-integration

|    Super VBMeta   | (variable-length)
|Backup Super VBMeta| (variable-length)

The Super VBMeta records the physical offset of each
logical partition's vbmeta within the /super partition.
The size of the super vbmeta is a multiple of 4K and
Primary/Backup is half and half.

Bug: 137054296
Test: m libvbmeta_test
Test: ./out/host/linux-x86/nativetest/libvbmeta_test/libvbmeta_test
Change-Id: I01aeadd850750ae87d9125484c1b1f570bb84756