Construct the vbmeta_table image
This commit constructs the vbmeta_table image to eliminate the
need of adding /vbmeta_system and/or /vbmeta_vendor when AVB
chain partition is used with Android Dynamic Partition.

See BOARD_AVB_VBMETA_SYSTEM under the link:
https://android.googlesource.com/platform/external/avb/#build-system-integration

|    VBMeta Table   | (fixed-length 2KiB)
|Backup VBMeta Table| (fixed-length 2KiB)

The VBMeta Table records the physical offset and the size of each
logical partition's vbmeta within the /super partition.

Bug: 137054296
Test: m libvbmeta_test
Test: ./out/host/linux-x86/nativetest/libvbmeta_test/libvbmeta_test
Change-Id: I01aeadd850750ae87d9125484c1b1f570bb84756