Update patch set 1
Patch Set 1:

Also, I wonder if it'd be better to have this function return the list as an output parameter then let the caller print them appropriately.  It'd be more generic, especially since we know that libmodprobe will be used in environments where std::cout is meaningless, such as in first_stage_init.

Not that we'd expect it to be called from there, but maybe for debugging, and having the vector would be better.

Plus, it'll allow it to be tested more easily, especially since there aren't any tests for this function in this current CL.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>