libutils: Introduce static String16
This is a backward compatible implementation of compile time
constructed String16 support.

As much as we'd like a regular constexpr constructor for String16, we
want to make sure the regular non-static String16 does not regress.
We also need to make sure prebuilts built with previous version of
String16 still works with new libutils.  This means we cannot change
the size of String16 objects and we cannot make anything virtual.

To add a flag to indicate whether a String16 is static without
increasing the size of non-static String16 objects, we repurpose a
reserved field in SharedBuffer as "for client use".  With this, we can
tag every String16 and perform memory operation differently based on
how the underlying buffers are allocated.

While we want to make the usage syntax simple, the fact that we cannot
change String16 size means we need to have an extra expression for
allocating the static storage.  The end result is something like:

    constexpr auto mystr_data = String16::makeStaticData(u"foo");
    const String16 mystr(mystr_data);

By doing so, we are able to eliminate the runtime construction of a
String16 and move it out of .bss section.

Bug: 138856262
Test: Run newly added unit tests.
Change-Id: I72bb8dc27a59b9ef34e0d934bc1e00b0f675855a