Update patch set 1
Patch Set 1:

Thanks. Isn't it possible to write a static function that wraps the statically-allocated block passing? 

I also thought a bit about tagging the data pointer so we can tell the difference between a static data block and a regular one at compile time without the reserved field check, but I don't think we can do that in a constexpr function, since we can't get at the bits of a pointer.

What I think would be even better would be encoding the location of the static buffer relative to the location of the String16 so that we don't even need a relocation. I don't see a way of getting there from pure C++ without linker help though.

Patch-set: 1
CC: Gerrit User 1130394 <1130394@85c56323-6fa9-3386-8a01-6480fb634889>