libsysutils: Remove an unnecessary bss object
The way kMaxSingleAddressLength is defined now requires it to be
calculated the first time it is used.  The calculated value is then
stored in the .bss section, which incurs a memory cost of 4KB.  Let's
make it a constexpr expression so that it is calculated at build time.
Since this is the only object in the .bss section in libsysutils,
removing it gets rid of .bss section entirely.

Bug: 138856262
Test: Build libsysutils on cuttlefish and check its .bss section is
      gone.

Change-Id: I0224db33051258d943f56df08ebd8fba32a555a9