adb: init: fs_mgr: Add adb-remount-sh to TEST_MAPPING
Any changes in adb, init or fs_mgr can affect the success or failure
of the adb remount functionality.  Trigger a pre-submit when any of
these directories are touched.  We limit to subset of directories
because this test runs for about 5 minutes on a local cuttlefish
instance and changes elsewhere generally do not impact adb remount
operations and could report false failures if adb, init or fs_mgr
is currently in a bad state.

The test is minimal because of inherent limitations on cuttlefish
(missing fastboot support, verity disabled) and comprise a few
reboots to initialize, setup, push test content, revert test content,
revert device to initial state.  This test can not be sharded with
any other tests, and failures during this test can result in the
device being in an indeterminate state.

Test: adb-remount-test.sh on local cuttlefish instance
Bug: 138649540
Change-Id: Ifa57a61ececbe500aaf5dbd7680c364c33e61ef5