Update patch set 1
Patch Set 1: Code-Review+1

Thanks for fixing this. I think this would have meant that if a user manually initiated atrace, we would have not restarted traced_probes afterwards (until a reboot). This is certainly something to be aware of but hopefully is not too common.

+Primiano FYI

Patch-set: 1
Reviewer: Gerrit User 1080238 <1080238@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1