Update patch set 3
Patch Set 3:

How hard is it to move to genfscon?

My big hesitation here is that genfscon is faster and more accurate than doing labeling in userspace, whereas taking this CL adds more complexity.  It seems like a win-win to move everything to genfscon and not add in this complexity.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>