Update patch set 5
Patch Set 5:

> Does the class actually add much value here?  We could have a std::string temp_dir and if it's not-empty do the rename / rmdir actions when needed.

The main purpose of the class is to rmdir on return with RAII. If there's a shorter way to do that I'm all for it. I need to be able to set it up sometime after the variable is declared, and only in some code paths, and cancel it if all goes well.

All the nits seem good to me, will sort out tomorrow!

Patch-set: 5
Reviewer: Gerrit User 1060907 <1060907@85c56323-6fa9-3386-8a01-6480fb634889>