Update patch set 5
Patch Set 5:

> Can you use the existing TemporaryDirectory?

That was my first approach, but it involved such drastic changes to TemporaryDirectory that I didn't think it was worth it. The biggest difficulty was that we create the variable early, then populate it with a location later if a temporary directory is needed. TemporaryDirectory's no-arguments constructor immediately creates the directory, and lots of things depend on that.

Patch-set: 5
Reviewer: Gerrit User 1060907 <1060907@85c56323-6fa9-3386-8a01-6480fb634889>