Update patch set 5
Patch Set 5:

Are we talking about android::base::TemporaryDir?  That also creates the directory in /data/local/tmp, whereas this creates it as the path that it should be with the random prefix on the end.  The latter seems to be better, but I guess it doesn't really matter since everything will be in /data in any case?

This all assumes that fscrypt_lookup_action() will only ever have an action for files in /data, right?  That's something we'll want to guarantee, since there is plenty of code that calls mkdir in /dev and other places where we probably don't want to do the rename.

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>