Update patch set 4
Patch Set 4:

We don't support || in init scripts.  We do allow checking for if a property is empty or not.  Assuming you use the abilist properties, you could do

on property:apexd.status=ready && property:ro.product.cpu.abilist32=*
  exec 32 bit check
on property:apexd.status=ready && property:ro.product.cpu.abilist64=*
  exec 64 bit check

I wrote on the doc, that maybe this is all better as a shell script.

Also, I'm only just now remembering that ExecWithRebootOnFailure() boots into recovery with a prompt to wipe data.  Is that what we'd want to do with this failure or should there be text specific to crypto if this test fails?  We may need to customize the string that we write to the bootloader.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>