Update patch set 2
Patch Set 2:

This was never intended to be used in production but rather to provide a way to audit syscall usage by the platform in internal testing. I haven't used it recently but it seems like a useful thing to have around, but if it's causing problems, it could be reapplied locally when doing the testing.

Patch-set: 2
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>