Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

Users prefer that reboot is not interrupted, and that the push to SD card file is destroyed is also expected, as for a friendly prompt, this is the user experience.
In the current implementation logic, I think /data/property/persistent_properties will be synchronized before umount. Currently sync is done before the kill service, which will increase the chance that the sync operation will take more time, if some services are doing a lot of io work.
So I don't think there is a need to do a sync operation before the kill service.
bug link : https://partnerissuetracker.corp.google.com/issues/140600729

Patch-set: 1
Reviewer: Gerrit User 1484529 <1484529@85c56323-6fa9-3386-8a01-6480fb634889>