Update patch set 8
Patch Set 8:

Ping. We believe all comments have been addressed. Nick, if the feature is turned off by default, is this still a security risk ? On our Android runtime, we would only enable that Unix listen socket when the user has "adb debugging on" in the Android settings and it would replace the tcp listening socket, which is a strict improvement and easier to secure inside Android.

Patch-set: 8
Reviewer: Gerrit User 1107893 <1107893@85c56323-6fa9-3386-8a01-6480fb634889>