Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Adding nnk@ and smoreland@ too.
> > 
> > I don't think there's much update.  We had an internal bug complaining about this, since the visibility wasn't good for people running 'logcat' to find out why their service didn't launch, but now that we're including the kernel log in logcat by default, we closed that bug as fixed.
> 
> Yeah, we've talked about doing this type of change for a long time, but the typical feedback is that thinking about domains early is important. For instance, without this check in place, audits are extremely useless.

yep, exactly. The goal of this check is to force the separation of init services preemptively, even if each domain is in permissive mode ala  userdebug_or_eng(`permissive foo;') . Not doing this misattributes all SELinux denials to the init domain, which just makes it harder later.

Patch-set: 2