adb: split host USB writes into 16k chunks.
We've seen USB writes failing due to inability to allocate contiguous
chunks of memory in the kernel on devices, but it looks like the same
problem can occur on the host, as well. Reduce our write sizes on the
host to a size (<= PAGE_ALLOC_COSTLY_ORDER) that always succeeds.

Test: not yet
Change-Id: I1517c348375b829dfff6796c4e9d394802b02d5b