WIP boringssl_self_test: make failure case more debuggable.
Run the self tests a second time once logd is running, so
that in the failure case one can obtain logs without needing
a serial connection (TODO: Is this the case?).

The initial invocation is run with "exec", so that the device
attempts to continue booting even if the invocation failed.

This is particularly helpful when the invocation failed for
a reason that wouldn't affect the rest of the system (like
a hardware problem), for example if a bug in a linker script
stops boringssl_self_test loading some shared library.

Test: [not yet tested, this CL is only for discussion]
Bug: 137267623
Change-Id: Ief583769f365d5a2f63be2abac3e8e57e7c2acb1