Update patch set 1
Patch Set 1:

My point is "we should be able to do this without all this copy and paste littering the tree". Can't we generate all this boilerplate instead? It doesn't seem to be adding any value that would warrant checking it in.

Patch-set: 1
Tag: mailMessageId=<CAJgzZopmorbkXWuhqKAVZWe2LMQo+3_xxC8CwNcODajYYKzkmA@mail.gmail.com>