Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This CL is only for discussion for now; this is completed untested, I don't actually know whether this would have the desired effect of failure logs being included in logcat or similar. Do you happen to know whether it would?
> 
> At least they'd have a fighting chance to make it to logd. However, I suspect the reboot happening immediately afterwards will stop them from being shipped (via adbd?) to "adb logcat" on the host.
> 
> There's also the problem with confusion from the SELinux denials when the first run fails, which are explained by errors much later in the log.
> 
> A better approach would be to ensure that logcat messages do get captured even before logd starts. Tom, can't liblog send them to the kernel log if it cannot get a connection to logd, at least during boot?

I was also figuring that kernel logging would be a good plan here.  The real question we had was whether or not TH would be able to get the logcat messages, since it wouldn't get the kernel logging for sure (except for cuttlefish, in which case we'd always get the kernel log...)

liblog can, but probably best to not use that right now since I'm still pending rewriting that.  Does this even use liblog now though?  It seems that it's all logging to stderr.  There's `logwrapper` that can wrap programs that log to stdout/stderr and redirect to kernel/logd/etc that could be handy here.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>