Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The "CANNOT LINK EXECUTABLE" error goes to both stderr and logcat. However, most of the verbose linker debug I usually enable to figure out linker config bugs only goes to logcat. My workflow in these cases is usually to reproduce locally and then enable linker debug with e.g. http://r.android.com/1013691. Sometimes I do the same through TH.
> 
> I see two approaches that complement each other:
> 
> 1. Getting kernel log coverage during boot. Piping process stderr's in there would be nice, but logcat messages are generally much more comprehensive in my experience, so if those were piped through to the kernel log by default, no ifs and buts (before logd takes over), then it'd be great. Sure, it requires either Cuttlefish or local serial cable, which sure is a bit of a pain but still really useful as a baseline with good coverage.
>
> 2. Improve logcat log shipping by a) starting logd (and adbd?) as early as possible, and b) somehow delay reboot until the logcat has been sent off. That would improve the debugging experience both in TH and locally. I reckon there will always be a point when this is no longer feasible and #1 comes into play.

We realistically shouldn't have many services starting before logd.  Of those that do, most already redirect their logs to dmesg (ueventd, apex boot streap, init).  This may be yet another service in that category, but it's not something that we should encourage.

I'm not sure about delaying the reboot.  I talked earlier with jmgao@ about starting adbd earlier and he did not want to do that.

> 
> > liblog can, but probably best to not use that right now since I'm still pending rewriting that.
> 
> Cool, is kernel logging fallback part of that work? Is there a ticket tracking this?

The same bug for merging of libbase and liblog would be it.

> 
> >  There's `logwrapper` that can wrap programs that log to stdout/stderr and redirect to kernel/logd/etc that could be handy here.
> 
> Thanks for the tip. Even if it only catches the "CANNOT LINK" message it'd be an improvement and probably a better option than the double run in this CL.

I really think that the best idea to run this with logwrapper.  We'll get the full logs in cuttlefish's serial output.  I don't think we're going to see a significant number of failures and I think it'll be even less likely that we'll start seeing failures on specific platforms but not others (and not cuttlefish).

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>