Update patch set 2
Patch Set 2:

(1 comment)

> I'm not sure about delaying the reboot.  I talked earlier with jmgao@ about starting adbd earlier and he did not want to do that.

Fair enough, I didn't expect any low-hanging fruit there. That's why I think the fallback to kernel log before logd is ready is the main take-away here.

> The same bug for merging of libbase and liblog would be it.

You mean http://b/119867234? Didn't see the connection there, but ok. (Btw, that ticket is rather the opposite of merging libbase and liblog, the key thing there is to get a libbase that support duplicate loads.)

> /.../ even less likely that we'll start seeing failures on specific platforms but not others (and not cuttlefish).

In the particular case that gave rise to this it was a problem in hwasan builds, which rules out Cuttlefish. We shouldn't draw too hard conclusions from this, but hwasan has a fair share of peculiarities which I suspect will result in more platform-specific bugs.

Patch-set: 2
CC: Gerrit User 1005967 <1005967@85c56323-6fa9-3386-8a01-6480fb634889>