Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> > I'm not sure about delaying the reboot.  I talked earlier with jmgao@ about starting adbd earlier and he did not want to do that.
> 
> Fair enough, I didn't expect any low-hanging fruit there. That's why I think the fallback to kernel log before logd is ready is the main take-away here.
> 
> > The same bug for merging of libbase and liblog would be it.
> 
> You mean http://b/119867234? Didn't see the connection there, but ok. (Btw, that ticket is rather the opposite of merging libbase and liblog, the key thing there is to get a libbase that support duplicate loads.)

Yes, that one.  Eventually it'll end up being a merging of the logging from both.

> 
> > /.../ even less likely that we'll start seeing failures on specific platforms but not others (and not cuttlefish).
> 
> In the particular case that gave rise to this it was a problem in hwasan builds, which rules out Cuttlefish. We shouldn't draw too hard conclusions from this, but hwasan has a fair share of peculiarities which I suspect will result in more platform-specific bugs.

Got it.  Honestly though, I'm doubting more and more that this is really needed.  hwasan is a brand new configuration in TH and mixing it with apex is a complex situation.  I'm not too sure how often we're going to see that again.  I don't think it was extremely difficult to debug since we have multiple hwasan targets that we could have used with debug cables.  I see that this is already abandoned and I'd figure let's keep it like that and wait to see if we do see difficult to debug issues arising in the future.

Patch-set: 2