Temporarily don't reboot if self test binaries are missing.
After OTAs, the system might be running with an older version
of the Conscrypt APEX that does not yet have the
boringssl_self_test{32,64} binaries.

As a short-term workaround, this CL changes init.rc to tolerate
this (not reboot into bootloader) by replacing the
exec_reboot_on_failure invocations of the binaries from the apex
with exec. If the ones in /system fail, then we still reboot.

This CL should be reverted once we have a more long term fix.

Bug: 140996463
Test: Treehugger

Change-Id: I700aaeb852c1436782864ec796e3b9d30bc4cde6