Update patch set 2
Patch Set 2:

(1 comment)

> Yes, that one.  Eventually it'll end up being a merging of the logging from both.

Sounds good, as long as the problematic global state gets out of libbase.

> Got it.  Honestly though, I'm doubting more and more that this is really needed.

I assume you're talking about boringssl_self_test specifically and not the kernel log fallback.

> we have multiple hwasan targets that we could have used with debug cables.

Did that, but to be clear there were neither logd nor stderr output in the serial log. I was only able to see what was happening by removing the exec_reboot_on_failure and run it later from the shell, but I wouldn't count in that being a possibility for init invocations in general.

Btw, wouldn't it be a good idea if init also piped stderr's to the kernel log by default, so that people couldn't do the mistake of leaving out logwrapper?

Patch-set: 2